# 🩹 Fix: Optimize Cache middleware handler