## Description

This PR optimize cache middleware's performace, with the following changes:
* Use Go's `slices.Contains` to replace original for loop (This should not increase performance but only reduce code)
* Use `strings.Builder` for concat strings for `key`
* `requestMethod := c.Method()` to avoid call `Method()` multiple times

## Benchmark:
before
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^Benchmark_Cache$ github.com/gofiber/fiber/v3/middleware/cache

goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v3/middleware/cache
cpu: AMD Ryzen 7 PRO 5850U with Radeon Graphics
=== RUN   Benchmark_Cache
Benchmark_Cache
Benchmark_Cache-16       4430882               271.5 ns/op            16 B/op          2 allocs/op
PASS
ok      github.com/gofiber/fiber/v3/middleware/cache    1.485s
```

after
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^Benchmark_Cache$ github.com/gofiber/fiber/v3/middleware/cache

goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v3/middleware/cache
cpu: AMD Ryzen 7 PRO 5850U with Radeon Graphics
=== RUN   Benchmark_Cache
Benchmark_Cache
Benchmark_Cache-16       3884007               303.4 ns/op            32 B/op          3 allocs/op
PASS
ok      github.com/gofiber/fiber/v3/middleware/cache    1.498s
```

## Changes introduced

List the new features or adjustments introduced in this pull request. Provide details on benchmarks, documentation updates, changelog entries, and if applicable, the migration guide.

- [x] Benchmarks: Describe any performance benchmarks and improvements related to the changes.
- [ ] Documentation Update: Detail the updates made to the documentation and links to the changed files.
- [ ] Changelog/What's New: Include a summary of the additions for the upcoming release notes.
- [ ] Migration Guide: If necessary, provide a guide or steps for users to migrate their existing code to accommodate these changes.
- [ ] API Alignment with Express: Explain how the changes align with the Express API.
- [ ] API Longevity: Discuss the steps taken to ensure that the new or updated APIs are consistent and not prone to breaking changes.
- [ ] Examples: Provide examples demonstrating the new features or changes in action.

## Type of change

Please delete options that are not relevant.

- [ ] New feature (non-breaking change which adds functionality)
- [ ] Enhancement (improvement to existing features and functionality)
- [ ] Documentation update (changes to documentation)
- [ ] Performance improvement (non-breaking change which improves efficiency)
- [x] Code consistency (non-breaking change which improves code reliability and robustness)

## Checklist

Before you submit your pull request, please make sure you meet these requirements:

- [x] Followed the inspiration of the Express.js framework for new functionalities, making them similar in usage.
- [x] Conducted a self-review of the code and provided comments for complex or critical parts.
- [ ] Updated the documentation in the `/docs/` directory for [Fiber's documentation](https://docs.gofiber.io/).
- [ ] Added or updated unit tests to validate the effectiveness of the changes or new features.
- [x] Ensured that new and existing unit tests pass locally with the changes.
- [ ] Verified that any new dependencies are essential and have been agreed upon by the maintainers/community.
- [x] Aimed for optimal performance with minimal allocations in the new code.
- [x] Provided benchmarks for the new code to analyze and improve upon.

## Commit formatting

Please use emojis in commit messages for an easy way to identify the purpose or intention of a commit. Check out the emoji cheatsheet here: [CONTRIBUTING.md](https://github.com/gofiber/fiber/blob/master/.github/CONTRIBUTING.md#pull-requests-or-commits)