@ReneWerner87 2024-06-11
@orvillesimba can you please make another benchmark and post before and after

in the upper one the numbers have gotten worse
it has increased from 2 to 3 allocations and the time from 271ns to 303ns